+++ src/wechaty-mixins/puppet-mixin.ts
@@ -25,7 +25,8 @@ import type {
 
 import type { GErrorMixin } from './gerror-mixin.js'
 import type { IoMixin }     from './io-mixin.js'
-import { diffPayload, InfoUpdateInterface } from '../schemas/update.js'
+import type { InfoUpdateInterface } from '../schemas/update.js'
+import { diffPayload } from '../pure-functions/update.js'
 
 const PUPPET_MEMORY_NAME = 'puppet'
 
@@ -442,14 +443,12 @@ const puppetMixin = <MixinBase extends WechatifyUserModuleMixin & GErrorMixin & 
                     const oldPayload = JSON.parse(JSON.stringify(contact?.payload || {}))
                     await contact?.ready(true)
                     const newPayload = JSON.parse(JSON.stringify(contact?.payload || {}))
-                    const difference = diffPayload(oldPayload, newPayload)
                     const updateEvent: InfoUpdateInterface = {
                       type: payloadType,
                       id: payloadId,
-                      fresh: difference.fresh,
-                      updates: difference.updates,
+                      updates: diffPayload(oldPayload, newPayload),
                     }
-                    if (updateEvent.fresh || updateEvent.updates.length > 0) {
+                    if (updateEvent.updates.length > 0) {
                       this.emit('update', updateEvent)
                       contact?.emit('update', updateEvent)
                     }
@@ -460,14 +459,12 @@ const puppetMixin = <MixinBase extends WechatifyUserModuleMixin & GErrorMixin & 
                     const oldPayload = JSON.parse(JSON.stringify(room?.payload || {}))
                     await room?.ready(true)
                     const newPayload = JSON.parse(JSON.stringify(room?.payload || {}))
-                    const difference = diffPayload(oldPayload, newPayload)
                     const updateEvent: InfoUpdateInterface = {
                       type: payloadType,
                       id: payloadId,
-                      fresh: difference.fresh,
-                      updates: difference.updates,
+                      updates: diffPayload(oldPayload, newPayload),
                     }
-                    if (updateEvent.fresh || updateEvent.updates.length > 0) {
+                    if (updateEvent.updates.length > 0) {
                       this.emit('update', updateEvent)
                       room?.emit('update', updateEvent)
                     }
